# dcapal-backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FeeStructure": {
        "oneOf": [
          {
            "properties": {
              "type": {
                "enum": [
                  "zeroFee"
                ],
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "properties": {
              "feeAmount": {
                "type": "string"
              },
              "type": {
                "enum": [
                  "fixed"
                ],
                "type": "string"
              }
            },
            "required": [
              "feeAmount",
              "type"
            ],
            "type": "object"
          },
          {
            "properties": {
              "feeRate": {
                "type": "string"
              },
              "maxFee": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "minFee": {
                "type": "string"
              },
              "type": {
                "enum": [
                  "variable"
                ],
                "type": "string"
              }
            },
            "required": [
              "feeRate",
              "minFee",
              "type"
            ],
            "type": "object"
          }
        ]
      },
...
      "PortfolioAssetRequest": {
        "properties": {
          "aclass": {
            "type": "string"
          },
          "averageBuyPrice": {
            "type": "string"
          },
          "baseCcy": {
            "type": "string"
          },
          "fees": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionFeesRequest"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "qty": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "targetWeight": {
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "aclass",
          "baseCcy",
          "provider",
          "qty",
          "targetWeight",
          "price",
          "averageBuyPrice"
        ],
        "type": "object"
      },
      "PortfolioAssetResponse": {
        "description": "An asset held by a portfolio.",
        "properties": {
          "aclass": {
            "description": "The asset class.",
            "type": "string"
          },
          "averageBuyPrice": {
            "description": "The average price paid for the holding.",
            "type": "string"
          },
          "baseCcy": {
            "description": "The asset currency.",
            "type": "string"
          },
          "fees": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionFeesResponse",
                "description": "The asset-level transaction fee settings."
              }
            ]
          },
          "name": {
            "description": "The asset display name.",
            "type": "string"
          },
          "price": {
            "description": "The latest known price.",
            "type": "string"
          },
          "provider": {
            "description": "The data provider for the asset.",
            "type": "string"
          },
          "qty": {
            "description": "The quantity held.",
            "type": "string"
          },
          "symbol": {
            "description": "The provider symbol for the asset.",
            "type": "string"
          },
          "targetWeight": {
            "description": "The target portfolio weight.",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "aclass",
          "baseCcy",
          "provider",
          "qty",
          "targetWeight",
          "price",
          "averageBuyPrice"
        ],
        "type": "object"
      },
      "PortfolioRequest": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/PortfolioAssetRequest"
            },
            "type": "array"
          },
          "fees": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionFeesRequest"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lastUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quoteCcy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "quoteCcy",
          "assets",
          "lastUpdatedAt"
        ],
        "type": "object"
      },
      "PortfolioResponse": {
        "description": "A portfolio returned by the synchronization endpoint.",
        "properties": {
          "assets": {
            "description": "The assets held by the portfolio.",
            "items": {
              "$ref": "#/components/schemas/PortfolioAssetResponse"
            },
            "type": "array"
          },
          "fees": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionFeesResponse",
                "description": "The portfolio-level transaction fee settings."
              }
            ]
          },
          "id": {
            "description": "The portfolio identifier.",
            "format": "uuid",
            "type": "string"
          },
          "lastUpdatedAt": {
            "description": "The timestamp used to compare this state with a client copy.",
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "description": "The portfolio display name.",
            "type": "string"
          },
          "quoteCcy": {
            "description": "The portfolio quote currency.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "quoteCcy",
          "assets",
          "lastUpdatedAt"
        ],
        "type": "object"
      },
...
      "SyncPortfoliosRequest": {
        "properties": {
          "deletedPortfolios": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "portfolios": {
            "items": {
              "$ref": "#/components/schemas/PortfolioRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "portfolios",
          "deletedPortfolios"
        ],
        "type": "object"
      },
      "SyncPortfoliosResponse": {
        "description": "The result of synchronizing the client's portfolio set.",
        "properties": {
          "deletedPortfolios": {
            "description": "Portfolio identifiers that the client should remove.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "updatedPortfolios": {
            "description": "Portfolios whose server state should be applied by the client.",
            "items": {
              "$ref": "#/components/schemas/PortfolioResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "updatedPortfolios",
          "deletedPortfolios"
        ],
        "type": "object"
      },
      "TransactionFeesRequest": {
        "properties": {
          "feeStructure": {
            "$ref": "#/components/schemas/FeeStructure"
          },
          "maxFeeImpact": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "feeStructure"
        ],
        "type": "object"
      },
      "TransactionFeesResponse": {
        "description": "Transaction fee settings returned by the API.",
        "properties": {
          "feeStructure": {
            "$ref": "#/components/schemas/FeeStructure",
            "description": "The fee calculation model."
          },
          "maxFeeImpact": {
            "description": "The maximum allowed fee impact, when configured.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "feeStructure"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DCA-Pal APIs",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sync/portfolios": {
      "post": {
        "operationId": "sync_portfolios",
        "parameters": [
          {
            "description": "Bearer JWT token",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPortfoliosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPortfoliosResponse"
                }
              }
            },
            "description": "Portfolios synchronized"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    }
  }
}