# packages/api-client/src/gen-mocks/model/portfolioAssetRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { TransactionFeesRequest } from './transactionFeesRequest';

export interface PortfolioAssetRequest {
  aclass: string;
  averageBuyPrice: string;
  baseCcy: string;
  fees?: null | TransactionFeesRequest;
  name: string;
  price: string;
  provider: string;
  qty: string;
  symbol: string;
  targetWeight: string;
}