# packages/api-client/src/gen-mocks/model/portfolioRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { PortfolioAssetRequest } from './portfolioAssetRequest';
import type { TransactionFeesRequest } from './transactionFeesRequest';

export interface PortfolioRequest {
  assets: PortfolioAssetRequest[];
  fees?: null | TransactionFeesRequest;
  id: string;
  lastUpdatedAt: string;
  name: string;
  quoteCcy: string;
}