# packages/api-client/src/gen-mocks/model/syncPortfoliosResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { PortfolioResponse } from './portfolioResponse';

export interface SyncPortfoliosResponse {
  deletedPortfolios: string[];
  updatedPortfolios: PortfolioResponse[];
}