# packages/api-client/src/gen-mocks/model/transactionFeesResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { FeeStructure } from './feeStructure';

export interface TransactionFeesResponse {
  feeStructure: FeeStructure;
  /** @nullable */
  maxFeeImpact?: string | null;
}