# packages/api-client/src/gen/model/feeStructure.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */

export type FeeStructure = {
  type: 'zeroFee';
} | {
  feeAmount: string;
  type: 'fixed';
} | {
  feeRate: string;
  /** @nullable */
  maxFee?: string | null;
  minFee: string;
  type: 'variable';
};