# packages/api-client/src/gen/model/portfolioAssetResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { TransactionFeesResponse } from './transactionFeesResponse';

export interface PortfolioAssetResponse {
  aclass: string;
  averageBuyPrice: string;
  baseCcy: string;
  fees?: null | TransactionFeesResponse;
  name: string;
  price: string;
  provider: string;
  qty: string;
  symbol: string;
  targetWeight: string;
}