# packages/api-client/src/gen/model/portfolioResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { PortfolioAssetResponse } from './portfolioAssetResponse';
import type { TransactionFeesResponse } from './transactionFeesResponse';

export interface PortfolioResponse {
  assets: PortfolioAssetResponse[];
  fees?: null | TransactionFeesResponse;
  id: string;
  lastUpdatedAt: string;
  name: string;
  quoteCcy: string;
}