# packages/api-client/src/gen/model/syncPortfoliosRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * DCA-Pal APIs
 * OpenAPI spec version: 0.9.0
 */
import type { PortfolioRequest } from './portfolioRequest';

export interface SyncPortfoliosRequest {
  deletedPortfolios: string[];
  portfolios: PortfolioRequest[];
}